Ensured logging is properly initialized in child processes
