+++ wordfence/cli/malwarescan/malwarescan.py
@@ -164,7 +164,8 @@ class MalwareScanSubcommand(Subcommand):
                 match_all=self.config.match_all,
                 pcre_options=self._get_pcre_options(),
                 allow_io_errors=self.config.allow_io_errors,
-                debug=self.config.debug
+                debug=self.config.debug,
+                logging_initializer=self.context.initialize_logging
             )
         if io_manager.should_read_stdin():
             options.path_source = io_manager.get_input_reader()
